refactor, comments
