+++ projects/vdk-plugins/vdk-impala/src/vdk/plugin/impala/templates/load/dimension/scd1/01-test-if-view-matches-target.sql
