+++ projects/vdk-plugins/vdk-impala/src/vdk/plugin/impala/templates/load/dimension/scd1/04-compute-stats.sql
