+++ projects/vdk-plugins/vdk-impala/src/vdk/plugin/impala/templates/load/dimension/scd2/00-dimension-scd2-definition.py
@@ -23,22 +23,10 @@ class SlowlyChangingDimensionType2(TemplateArgumentsValidator):
     TemplateParams = SlowlyChangingDimensionType2Params
 
     def __init__(self) -> None:
-        super().__init__(
-            template_name="load/dimension/scd2",
-            sql_files=[
-                "00-test-if-view-matches-target.sql",
-                "01-insert-into-target.sql",
-                "02-refresh.sql",
-                "03-compute-stats.sql",
-            ],
-            sql_files_platform_is_responsible=[
-                "02-refresh.sql",
-                "03-compute-stats.sql",
-            ],
-        )
+        super().__init__()
 
 
-def get_validated_arguments(job_input: IJobInput):
-    return SlowlyChangingDimensionType2().get_validated_args(
+def run(job_input: IJobInput):
+    SlowlyChangingDimensionType2().get_validated_args(
         job_input, job_input.get_arguments()
     )
