+++ projects/vdk-plugins/vdk-impala/src/vdk/plugin/impala/templates/load/dimension/scd2/02-insert-into-target.sql
