+++ projects/vdk-plugins/vdk-impala/src/vdk/plugin/impala/templates/load/fact/snapshot/01-test-if-view-matches-target.sql
