+++ projects/vdk-plugins/vdk-impala/src/vdk/plugin/impala/templates/load/fact/snapshot/04-compute-stats.sql
