--- projects/vdk-plugins/vdk-impala/src/vdk/plugin/impala/templates/load/validators/__init__.py
@@ -1,2 +1,0 @@
-# Copyright 2021 VMware, Inc.
-# SPDX-License-Identifier: Apache-2.0
