+++ projects/vdk-plugins/vdk-impala/src/vdk/plugin/impala/templates/load/versioned/01-test-if-view-matches-target.sql
