+++ projects/vdk-plugins/vdk-impala/src/vdk/plugin/impala/templates/load/versioned/02-insert-into-target.sql
