+++ projects/vdk-plugins/vdk-impala/src/vdk/plugin/impala/templates/load/versioned/03-refresh.sql
