+++ projects/vdk-plugins/vdk-impala/src/vdk/plugin/impala/templates/load/versioned/04-compute-stats.sql
