+++ projects/vdk-plugins/vdk-impala/tests/functional/jobs/load_dimension_scd1_template_job/02_run_load_dimension_scd1_template.py
@@ -1,7 +1,6 @@
 # Copyright 2021 VMware, Inc.
 # SPDX-License-Identifier: Apache-2.0
 from vdk.api.job_input import IJobInput
-from vdk.plugin.impala.templates.load.validators import dimension_scd1_definition
 
 __author__ = "VMware, Inc."
 __copyright__ = (
@@ -10,8 +9,7 @@ __copyright__ = (
 
 
 def run(job_input: IJobInput) -> None:
-    args = dimension_scd1_definition.get_validated_arguments(job_input)
     job_input.execute_template(
         template_name="load/dimension/scd1",
-        template_args=args,
+        template_args=job_input.get_arguments(),
     )
